+++ main.py
@@ -26,7 +26,6 @@ from util.mielib.custominput import bool_input
 from minecraft.install import Installer
 from util.cron import CronScheduler
 from util import configuration as c
-from util.backup import Backup
 from util.temp import PiTemp
 from util.syslog import log
 from util.date import Date
@@ -89,9 +88,10 @@ def parse(args):
         updateServer(update)
 
     if backup:
-        running_log.append('-bu {}'.format(c.Maintenance.backup_path))
+        running_log.append(f'-bu {c.Maintenance.backup_path}')
         cmd.runCommand("say System is backing up Minecraft world.")
-        filename = 'world.{}.zip'.format(Date.strippedTimestamp())
+        log('Backing up the current world...')
+        filename = f'world.{Date.strippedTimestamp()}.zip'
         Backup.put(Installer.server_dir, c.Maintenance.backup_path, filename)
 
     if method is not None:
@@ -184,9 +184,10 @@ def run_debug():
     # Shut off calling server commands for debugging purposes
     cmd.DEBUG = True
 
+    # Do NOT delete either of the DEBUGGING print statements!
     print('\n****** DEBUGGING STARTED ******\n')
-    # Implement any debug functionality below:
 
+    # Implement any debug functionality below:
     run()
 
     print('Waiting 2 minutes before continuing so server can start up properly')